LPS-177939 rename to align with the functionality
